+++ tests/capture_photos.py
