+++ tests/picam.py
