+++ tests/sensor_test.py
